+++ tests/test_simul_efuns/test_simul_efuns.cpp
@@ -143,7 +143,7 @@ TEST_F(SimulEfunsTest, callSimulEfun)
     current_object = simul_efun_ob;
     push_number(10);
     push_number(20);
-    call_simul_efun((unsigned short)index, 2);
+    call_simul_efun (index, 2);
     EXPECT_TRUE(sp->type == T_NUMBER) << "Return value type from simul efun 'sum' is not T_NUMBER.";
     EXPECT_EQ(sp->u.number, 30) << "Return value from simul efun 'sum' is not correct.";
     push_constant_string ("0");
@@ -152,7 +152,7 @@ TEST_F(SimulEfunsTest, callSimulEfun)
     func_name = findstring("concat");
     ASSERT_TRUE(func_name != nullptr) << "Failed to find string 'concat'.";
     index = find_simul_efun(func_name);
-    call_simul_efun((unsigned short)index, 2);
+    call_simul_efun (index, 2);
     EXPECT_TRUE(sp->type == T_STRING) << "Return value type from simul efun 'concat' is not T_STRING.";
     EXPECT_STREQ(sp->u.string, "300") << "Return value from simul efun 'concat' is not correct.";
